refactor(web): split TopPicksSection into topPicks modules
Extract settings UI into hooks and focused cards while keeping TopPicksSection as thin composition over the original behavior.
